mcp_server/pyproject.toml
[project]
name = "mcp-portfolio"
version = "0.1.0"
description = "MCP server providing read-only access to Portfolio Digital data via the FastAPI backend"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx",
    "python-dotenv",
    "matplotlib>=3.8",
]

[project.scripts]
mcp-portfolio = "mcp_portfolio:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_portfolio"]
